Allow changing JSON library
Many people prefer Jason over Poison because it can be faster. Phoenix
framework did the [same switch][1] over a year ago now.

This commit makes the json library customizable. It still defaults to
Poison to ensure that we don't break anything for people that are upgrading
the library.

It's now possible to use Jason (and even jiffy when using a wrapper).

The only other change is that encode function is not longer using
`pretty: true`. This should give a tiny performance boost.

[1]: https://github.com/phoenixframework/phoenix/commit/02a204a33dc57bc215bf19581f2256c6d40dffde