Add OpenTelemetry support
`opentelemetry` (sdk) dependency is marked for `test` only. We only need
`opentelemetry_api` to do the all instrumentation calls. The application
can then use `opentelemetry` or some other implementation if there will
be one in the future.

If the library user does not use `opentelemetry` then the OTEL calls are
basically no-ops.